Dialog System Fully Working
Changes:

There is now a trigger box at the beginning of the level that will trigger the beginning dialog of the level (where Mousey and Sir Bitsy meet). The sentence can be completed instantly if return is hit during the typewritter effect. Hitting return once the sentence is fully written moves onto the next sentence.

Removed temporary key in PlayerController to trigger dialog.

Added dialog detection script (used to detect then send the trigger to dialog system).

Notes:

Need to add remaining dialogs to json file and their respective triggers to the level.

Need to remove symbols in json file (issue with characters when converting the file to json).