Added Asynchronous Level Loading
Changes:

Added LoadingLevel scene where the next level is loaded in the background. My thought is we can have the movie play while the next level is loaded in this scene. I can always disable or remove this if needed.

Added EndOfLevelTrigger to MainGame scene to transition to the next scene once hit (can change if we want a victory screen or such before swapping scenes).